move /mnt/data mount check to utils
