+++ tools/flashy/lib/utils/system.go
@@ -377,3 +377,19 @@ var GetMTDMapFromSpecifier = func(deviceSpecifier string) (map[string]string, er
 	}
 	return mtdMap, nil
 }
+
+// IsDataPartitionMounted checks if /mnt/data is mounted
+var IsDataPartitionMounted = func() (bool, error) {
+	procMountsDat, err := fileutils.ReadFile(procMountsPath)
+	if err != nil {
+		return false, errors.Errorf("Cannot read /proc/mounts: %v", err)
+	}
+
+	regEx := `(?m)^[^ ]+ /mnt/data [^ ]+ [^ ]+ [0-9]+ [0-9]+$`
+	regExMap, err := GetAllRegexSubexpMap(regEx, string(procMountsDat))
+	if err != nil {
+		return false, errors.Errorf("regex error: %v", err)
+	}
+
+	return len(regExMap) != 0, nil
+}
